refactor 3

+++ include/quickjs/macros/types.h
@@ -13,7 +13,7 @@ typedef unsigned char bool;
 #endif
 
 typedef float       float32_t;
-typedef long double float64_t;
+typedef double      float64_t;
 
 #ifndef YES
 #define YES  1
